chore(bom): moved out old bom "fromEvent" remaining actions
